Moved ranking interface to model
