+++ src/Indexing/Index.ts
@@ -6,7 +6,7 @@ import IIndexTokenMeta from "./IIndexTokenMeta"
 import { ObjectLiteral } from "../Utility/JsonUtility"
 import IIndexOptions from "./IIndexOptions"
 import { WORD } from "../Tokenizer/TokenizerStrategy"
-import IRanker from "./Ranking/IRanker"
+import IRanker from "../Model/IRanker"
 import { TFIDF } from "./Ranking/rankingStrategy"
 import { IIndexStatistics } from "./IIndexStatistics"
 import IIndexDocument from "../Model/IIndexDocument"
